backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-groq
langchain-community
chromadb
tavily-python
yfinance
pandas
pydantic
python-dotenv
numpy
scipy
PyPortfolioOpt
